refactor: replace web xstate import with main entry
