--- src/types/xstate.d.ts
@@ -1,3 +1,0 @@
-declare module 'xstate/dist/xstate.web.js' {
-  export * from 'xstate';
-}
